move service registration to services.py
